Fix Project Structure
