+++ Teste/src/engine/Core.java
@@ -1,4 +1,4 @@
-package poo.game.project;
+package engine;
 
 public class Core {
 	public static void main (String[] args) {
