--- Teste/src/entity/Entity.java
@@ -1,6 +1,0 @@
-package entity;
-
-public class Entity {
-	private int x, y;
-	private int speed;
-}
