+++ Teste/src/game/Entity.java
@@ -1,0 +1,6 @@
+package game;
+
+public class Entity {
+	private int x, y;
+	private int speed;
+}
